ai-context/requirements.txt
google-generativeai==0.8.5
python-dotenv==1.0.0
pyyaml==6.0.1
